requirements.txt
# Core dependencies (Lightweight version for Railway)
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Image processing (minimal)
pillow==10.4.0
PyPDF2==3.0.1

# Hugging Face API client (no local models)
huggingface-hub==0.34.4

# Utilities
aiofiles==23.2.1
numpy>=1.24.0
requests==2.32.3

# Note: We're NOT installing torch, transformers, or other heavy ML libraries
# We'll use Hugging Face's Inference API instead
